pyproject.toml
[project]
name = "shiftedblog"
version = "0.1.3"
description = "Automated blogging tool"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "django>=6.0",
    "django-axes>=8.1.0",
    "django-ckeditor-5>=0.2.18",
    "django-ratelimit>=4.1.0",
    "django-redis>=5.4.0",
    "django-taggit>=6.1.0",
    "django-two-factor-auth>=1.18.0",
    "phonenumbers>=9.0.8",
    "pillow>=11.2.1",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.0",
    "qrcode>=7.4.2",
    "redis>=5.0.0",
    "requests>=2.32.5",
    "yadisk>=3.4.0",
]

[project.optional-dependencies]
dev = [
    "pyright>=1.1.0",
    "ruff>=0.8.0",
]

[tool.ruff]
target-version = "py314"
line-length = 88
exclude = [
    ".git",
    ".venv",
    "venv",
    "build",
    "dist",
    "*.egg-info",
    "**/migrations",
]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "F",      # Pyflakes
    "I",      # isort
    "N",      # pep8-naming
    "W",      # pycodestyle warnings
    "UP",     # pyupgrade
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "SIM",    # flake8-simplify
    "T20",    # flake8-print (no print in production code)
    "DJ",     # flake8-django
    "RUF",    # Ruff-specific rules
]
ignore = [
    "DJ001",  # avoid setattr
    "T201",   # print found (allow in manage/scripts if needed)
]
extend-per-file-ignores = { "manage.py" = ["T201"] }

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.pyright]
pythonVersion = "3.14"
typeCheckingMode = "basic"
include = ["blog", "shiftedblog", "manage.py", "templates"]
exclude = [
    ".git",
    ".venv",
    "venv",
    "build",
    "dist",
    "**/__pycache__",
    "**/*.egg-info",
    "blog/migrations",
]
# Use project venv for import resolution (run with venv active or set venvPath/venv)
venvPath = "."
venv = ".venv"
# "warning" so type check can run without venv (e.g. CI); set "true" with venv for strict import checking
reportMissingImports = "warning"
reportMissingTypeStubs = false
reportUnnecessaryTypeIgnoreComment = true
reportUnnecessaryCast = true
reportUnnecessaryComparison = true
# Django: resolve from project root
extraPaths = ["."]
